# TrillionBit-API-stage/TrillionBit-API-stage/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bitex application programming interface (API) allows our clients to access and control their accounts, using custom written software. <br/> With the inclusion of the new BTC/AED currency pair, we've added new endpoints for some API calls (info about specific calls below). The old endpoints are still available and have not changed. We strongly suggest however, that you refer to the API for future references.",
        "version": "1.0.0",
        "title": "Bitex",
        "termsOfService": "http://bitex.com",
        "contact": {
        "email": "admin@bitex.com"
    },
    "license": {
        "name": "Bitex API",
        "url": "http://bitex.com"
    }
    },
    "host": "api.bitex.com",
    "basePath": "/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User login.",
                "description": "User login",
                "operationId": "user.login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User login",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "user.login"
                        ]
                    }
                ]
            }
        },
        "/auth/register": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Register user.",
                "description": "Register user.",
                "operationId": "user.register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User register",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "user.register"
                        ]
                    }
                ]
            }
        },
...
        "/trading/order/pending": {
            "post": {
                "tags": [
                    "Trading"
                ],
                "summary": "Get order pending.",
                "description": "Get order pending",
                "operationId": "order.pending",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Get order pending",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user_id": {
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "market": {
                                    "type": "string"
                                },
                                "offset": {
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "limit": {
                                    "type": "integer",
                                    "format": "int64"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "create:order"
                        ]
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://api.bitex.com/api/auth/login",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
        "UserLogin": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "userDevice": {
                    "type": "object"
                }
            }
        },
        "UserRegister": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "example": "+91 9876543210"
                },
                "password": {
                    "type": "string"
                },
                "password2": {
                    "type": "string"
                },
                "country": {
                    "type": "string",
                    "example": "IN"
                }
            }
        },
...
    },
...
}